Reset HTTP/3 request on deadline exceeded
