+++ src/Grpc.AspNetCore.Server/Internal/GrpcProtocolConstants.cs
@@ -70,7 +70,10 @@ namespace Grpc.AspNetCore.Server.Internal
 #endif
 
         internal const string IdentityGrpcEncoding = "identity";
-        internal const int Http2ResetStreamNoError = 0;
+        internal const int Http2ResetStreamCancel = 0x8;
+#if NET6_0_OR_GREATER
+        internal const int Http3ResetStreamCancel = 0x010c;
+#endif
 
         internal static readonly HashSet<string> FilteredHeaders = new HashSet<string>(StringComparer.OrdinalIgnoreCase)
         {
@@ -114,6 +117,15 @@ namespace Grpc.AspNetCore.Server.Internal
         {
             return ReferenceEquals(encoding, IdentityGrpcEncoding) ||
                 string.Equals(encoding, IdentityGrpcEncoding, StringComparison.Ordinal);
+        }
+
+        internal static int GetCancelErrorCode(string protocol)
+        {
+#if NET6_0_OR_GREATER
+            return IsHttp3(protocol) ? Http3ResetStreamCancel : Http2ResetStreamCancel;
+#else
+            return Http2ResetStreamCancel;
+#endif
         }
     }
 }
