+++ src/Grpc.Net.Client/Grpc.Net.Client.csproj
@@ -32,6 +32,7 @@
     <Compile Include="..\Shared\NullableAttributes.cs" Link="Internal\NullableAttributes.cs" />
     <Compile Include="..\Shared\Http2ErrorCode.cs" Link="Internal\Http2ErrorCode.cs" />
     <Compile Include="..\Shared\Http3ErrorCode.cs" Link="Internal\Http3ErrorCode.cs" />
+    <Compile Include="..\Shared\HttpExceptionResolver.cs" Link="Internal\Http\HttpExceptionResolver.cs" />
   </ItemGroup>
 
   <ItemGroup>
