+++ test/FunctionalTests/Web/Server/DeadlineTests.cs
@@ -27,6 +27,7 @@ using Grpc.AspNetCore.FunctionalTests.Infrastructure;
 using Grpc.AspNetCore.Server.Internal;
 using Grpc.Core;
 using Grpc.Net.Client.Web;
+using Grpc.Shared;
 using Grpc.Tests.Shared;
 using NUnit.Framework;
 
@@ -102,12 +103,19 @@ namespace Grpc.AspNetCore.FunctionalTests.Web.Server
             httpRequest.Headers.Add(GrpcProtocolConstants.TimeoutHeader, "300m");
             httpRequest.Content = new GrpcStreamContent(requestStream);
 
-            // Act
-            var response = await grpcWebClient.SendAsync(httpRequest, HttpCompletionOption.ResponseHeadersRead).DefaultTimeout();
+            try
+            {
+                // Act
+                var response = await grpcWebClient.SendAsync(httpRequest, HttpCompletionOption.ResponseHeadersRead).DefaultTimeout();
 
-            // Assert
-            response.AssertIsSuccessfulGrpcRequest();
-            response.AssertTrailerStatus(StatusCode.DeadlineExceeded, "Deadline Exceeded");
+                // Assert
+                response.AssertIsSuccessfulGrpcRequest();
+                response.AssertTrailerStatus(StatusCode.DeadlineExceeded, "Deadline Exceeded");
+            }
+            catch (Exception ex) when (HttpExceptionResolver.ResolveRpcExceptionStatusCode(ex) == StatusCode.Cancelled)
+            {
+                // Ignore exception from deadline abort
+            }
         }
 
         [Test]
@@ -168,41 +176,48 @@ namespace Grpc.AspNetCore.FunctionalTests.Web.Server
             httpRequest.Headers.Add(GrpcProtocolConstants.TimeoutHeader, "200m");
             httpRequest.Content = new GrpcStreamContent(requestStream);
 
-            // Act
-            var grpcWebClient = CreateGrpcWebClient();
-            var response = await grpcWebClient.SendAsync(httpRequest, HttpCompletionOption.ResponseHeadersRead).DefaultTimeout();
-
-            // Assert
-            response.AssertIsSuccessfulGrpcRequest();
-
-            var responseStream = await response.Content.ReadAsStreamAsync().DefaultTimeout();
-            var pipeReader = PipeReader.Create(responseStream);
-
-            var messageCount = 0;
-
-            var readTask = Task.Run(async () =>
+            try
             {
-                while (true)
+                // Act
+                var grpcWebClient = CreateGrpcWebClient();
+                var response = await grpcWebClient.SendAsync(httpRequest, HttpCompletionOption.ResponseHeadersRead).DefaultTimeout();
+
+                // Assert
+                response.AssertIsSuccessfulGrpcRequest();
+
+                var responseStream = await response.Content.ReadAsStreamAsync().DefaultTimeout();
+                var pipeReader = PipeReader.Create(responseStream);
+
+                var messageCount = 0;
+
+                var readTask = Task.Run(async () =>
                 {
-                    var greeting = await MessageHelpers.AssertReadStreamMessageAsync<HelloReply>(pipeReader).DefaultTimeout();
-
-                    if (greeting != null)
+                    while (true)
                     {
-                        Assert.AreEqual($"How are you World? {messageCount}", greeting.Message);
-                        messageCount++;
+                        var greeting = await MessageHelpers.AssertReadStreamMessageAsync<HelloReply>(pipeReader).DefaultTimeout();
+
+                        if (greeting != null)
+                        {
+                            Assert.AreEqual($"How are you World? {messageCount}", greeting.Message);
+                            messageCount++;
+                        }
+                        else
+                        {
+                            break;
+                        }
                     }
-                    else
-                    {
-                        break;
-                    }
-                }
 
-            });
+                });
 
-            await readTask.DefaultTimeout();
+                await readTask.DefaultTimeout();
 
-            Assert.AreNotEqual(0, messageCount);
-            response.AssertTrailerStatus(StatusCode.DeadlineExceeded, "Deadline Exceeded");
+                Assert.AreNotEqual(0, messageCount);
+                response.AssertTrailerStatus(StatusCode.DeadlineExceeded, "Deadline Exceeded");
+            }
+            catch (Exception ex) when (HttpExceptionResolver.ResolveRpcExceptionStatusCode(ex) == StatusCode.Cancelled)
+            {
+                // Ignore exception from deadline abort
+            }
 
             // The server has completed the response but is still running
             // Allow time for the server to complete
